Fix: implicit conversion turns floating-point number into integer: 'double' to 'size_t'

+++ src/bloom_filter/basic.cpp
@@ -7,12 +7,12 @@ namespace bf {
 
 size_t basic_bloom_filter::m(double fp, size_t capacity) {
   auto ln2 = std::log(2);
-  return std::ceil(-(capacity * std::log(fp) / ln2 / ln2));
+  return static_cast<size_t>(std::ceil(-((double)capacity * std::log(fp) / ln2 / ln2)));
 }
 
 size_t basic_bloom_filter::k(size_t cells, size_t capacity) {
   auto frac = static_cast<double>(cells) / static_cast<double>(capacity);
-  return std::ceil(frac * std::log(2));
+  return static_cast<size_t>(std::ceil(frac * std::log(2)));
 }
 
 basic_bloom_filter::basic_bloom_filter(hasher h, size_t cells, bool partition)
